# SWATCH-2160: Increase SubscriptionDefinition lookup speed