--- a/swatch-product-configuration/build.gradle
+++ b/swatch-product-configuration/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation libraries["slf4j-api"]
     implementation libraries["jakarta-validation-api"]
     implementation libraries["guava"]
+    implementation libraries["caffeine"]
 
     testImplementation libraries["junit-jupiter"]
     testImplementation libraries["hamcrest-all"]