--- a/swatch-product-configuration/src/main/java/com/redhat/swatch/configuration/registry/SubscriptionDefinitionRegistry.java
+++ b/swatch-product-configuration/src/main/java/com/redhat/swatch/configuration/registry/SubscriptionDefinitionRegistry.java
@@ -49,6 +49,14 @@ public class SubscriptionDefinitionRegistry {
     return instance;
   }
 
+  /** Reset the registry instance for test purposes. */
+  public static synchronized void reset() {
+    SubscriptionDefinition.tagCache.invalidateAll();
+    SubscriptionDefinition.engIdCache.invalidateAll();
+    SubscriptionDefinition.roleCache.invalidateAll();
+    instance = new SubscriptionDefinitionRegistry();
+  }
+
   SubscriptionDefinitionRegistry() {
     subscriptions = new ArrayList<>();
     var options = new LoaderOptions();