@ben-manes 2024-09-09
> Schedule periodic cleanup so that cache maintenance occurs outside periods of heavy read usage.

The cache maintenance is very inexpensive and runs regularly to replay activity. However it is not a dedicated thread and defers to FJP.commonPool() when triggered. It is asynchronous because an eviction listener may be attached and foreign code has an unknown expense. Many like Infinispan use the calling thread, executor(Runnable::run).

The scheduler is for prompt expiration when the cache is idle, e.g. to close a websocket after a timeout. Some users need predictable cleanup/notification, whereas most don’t care. The GC is lazy anyway, so eagerly evicting an expired entry usually has no memory performance effects.

If your cache is fairly active then you don’t need the scheduler, but it does no harm. I’d also recommend using jmh for benchmarks to avoid jit warmup surprises.