@awood 2024-09-10
> I don't think we need a cache like this since the loaded data is coming from static files that won't change after 3 hours or other amount of time, so in practice, this is like having loaded this data in memory all the time.
> 

No, the entries will be evicted after 3 hours if they are not used.  Since nightly tallies (`BaseSnapshotRoller`) are what are doing the lookups, the cache should be empty for most of the day.

If we use weak references, the cache will be emptied much more frequently (degrading performance) since those values will be getting GCed in between tallies of the different orgs (since the values are only referenced by the cache itself).