@awood 2024-09-10
> If your cache is fairly active then you don’t need the scheduler, but it does no harm. I’d also recommend using jmh for benchmarks to avoid jit warmup surprises.

Thank you so much for commenting @ben-manes!  Our cache is going to be pretty inactive.  There will be just a few writes when it's empty and then a ton of reads during a nightly job.  After that it'll be inactive until the next night.  I'm not concerned so much about eviction occurring right away.  My concern was more that I didn't want to slow down cache activity while we are performing our burst of reads.  Does this seem like a good use-case for the scheduler?

Thanks again for stopping by our humble repo!