@ben-manes 2024-09-10
> My concern was more that I didn't want to slow down cache activity while we are performing our burst of reads.

That shouldn't be a problem. The reads are lock-free and shouldn't block, even under high load. We split the cache into its data and the policy, with a pub-sub model to keep the policy eventually consistent. This way you can get ConcurrentHashMap style performance and behavior (e.g per-entry computes), without being bottlenecked as we keep the policy up-to-date (global shared state, every read is a write). We use bounded buffers where when full on a read it drops events (as popularity skew will mean those items get added more often, so minimal impact to the hit rate) and the write buffer only get throttled when inserts greatly outpace evictions (no runaway memory leak). 

These are the jmh [benchmarks](https://github.com/ben-manes/caffeine/wiki/Benchmarks) run on my M3 Max laptop:

![image](https://github.com/user-attachments/assets/007edc5c-040b-4ae8-b121-f0ebfe8c9891)