@Sgitario 2024-09-12
> No, the entries will be evicted after 3 hours if they are not used. Since nightly tallies (`BaseSnapshotRoller`) are what are doing the lookups, the cache should be empty for most of the day.
> 
> If we use weak references, the cache will be emptied much more frequently (degrading performance) since those values will be getting GCed in between tallies of the different orgs (since the values are only referenced by the cache itself).

My main concern is about having extra data loaded in memory just in the tally process that is when we push the service memory's constraints more. For example: if the tally takes more than 3 hours, the cache will be refreshed several times which is not a problem, but the memory will keep this data longer. 

If this extra data is only a few MB, then that's ok. Otherwise, I would like that the garbage collector can free this data via either weak reference or maybe we can configure caffeine for doing this, even if this degradates the performance.