@ben-manes 2024-09-12
In that case using the scheduler is appropriate as it will trigger a cleanUp() when the entry expires. You can specify weak values but it might be collected prematurely, so I prefer that for lifecycle management rather than size control.