@awood 2024-09-12
> My main concern is about having extra data loaded in memory just in the tally process that is when we push the service memory's constraints more. For example: if the tally takes more than 3 hours, the cache will be refreshed several times which is not a problem, but the memory will keep this data longer.
> 
> If this extra data is only a few MB, then that's ok. Otherwise, I would like that the garbage collector can free this data via either weak reference or maybe we can configure caffeine for doing this, even if this degradates the performance.

These data structures are not large at all.  Our tallies take so long, that I would rather optimize for CPU usage than memory use.  There are 22 SubscriptionDefinitions and according to the profiler that totals to about 1KB of memory used.  Caffeine classes add about 4KB.  We can use the weak references approach, but it will result in substantially more cache evictions since every tally basically will have to repopulate whatever `Optional<SubscriptionDefinition>` objects got GCed.  Personally, I think that is worth the extra 5KB of extra memory use to improve the tally speed as much as possible.

Like @ben-manes mentioned, I can put the scheduler back in to trigger cleanUp's once the entries expire, but again that's extra CPU use for the sake of 5KB of memory.